requirements.txt
streamlit==1.40.0
psycopg2-binary==2.9.10
bcrypt==4.2.0
python-dotenv==1.0.1